+++ bid_stalker/site/facebook/__init__.py
@@ -1,0 +1,4 @@
+from .site import marketplace
+
+
+__all__ = [ 'marketplace' ]