+++ HISTORY.rst
@@ -2,7 +2,16 @@
 History
 =======
 
+******************
+1.0.0 (2026-03-09)
+******************
+
+* release for bidspotter
+* can read audiction by country
+* scan articles on audiction and send the data to elasticsearch
+
+******************
 0.0.1 (2026-03-03)
-------------------
+******************
 
 * First release on PyPI.